+++ src/components/field/FieldSection.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from '@dhis2/prop-types'
-import styles from './Fied.module.css'
+import styles from './Field.module.css'
 
 export const FieldSection = props => (
     <div className={styles.row}>{props.children}</div>
